+++ tests/dummy/app/Models/Post.php
@@ -23,11 +23,13 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Post extends Model
 {
 
     use HasFactory;
+    use SoftDeletes;
 
     /**
      * @var string[]