+++ tests/dummy/tests/Api/V1/Serializer.php
@@ -44,6 +44,7 @@ class Serializer
             'attributes' => [
                 'content' => $post->content,
                 'createdAt' => $post->created_at->jsonSerialize(),
+                'deletedAt' => optional($post->deleted_at)->jsonSerialize(),
                 'publishedAt' => optional($post->published_at)->jsonSerialize(),
                 'slug' => $post->slug,
                 'synopsis' => $post->synopsis,