+++ tests/dummy/app/JsonApi/V1/Posts/PostResource.php
@@ -36,6 +36,7 @@ class PostResource extends JsonApiResource
         return [
             'content' => $this->content,
             'createdAt' => $this->created_at,
+            'deletedAt' => $this->deleted_at,
             'publishedAt' => $this->published_at,
             'slug' => $this->slug,
             'synopsis' => $this->synopsis,
@@ -55,6 +56,8 @@ class PostResource extends JsonApiResource
         return [
             $this->relation('author')->showDataIfLoaded(),
             $this->relation('comments'),
+            $this->relation('media')
+                ->withData(fn() => $this->schema->relationship('media')->value($this->resource)),
             $this->relation('tags')->showDataIfLoaded(),
         ];
     }