+++ tests/dummy/app/Models/Image.php
@@ -21,11 +21,9 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Support\Str;
 
-class Video extends Model
+class Image extends Model
 {
 
     use HasFactory;
@@ -48,7 +46,7 @@ class Video extends Model
     /**
      * @var string[]
      */
-    protected $fillable = ['title', 'url'];
+    protected $fillable = ['url'];
 
     /**
      * @inheritDoc
@@ -60,21 +58,5 @@ class Video extends Model
         self::creating(static function (self $model) {
             $model->{$model->getKeyName()} = $model->{$model->getKeyName()} ?? Str::uuid()->toString();
         });
-    }
-
-    /**
-     * @return BelongsTo
-     */
-    public function owner(): BelongsTo
-    {
-        return $this->belongsTo(User::class);
-    }
-
-    /**
-     * @return MorphToMany
-     */
-    public function tags(): MorphToMany
-    {
-        return $this->morphToMany(Tag::class, 'taggable');
     }
 }