+++ tests/dummy/app/Models/Post.php
@@ -21,13 +21,16 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Post extends Model
 {
 
     use HasFactory;
+    use SoftDeletes;
 
     /**
      * @var string[]
@@ -64,11 +67,27 @@ class Post extends Model
     }
 
     /**
+     * @return BelongsToMany
+     */
+    public function images(): BelongsToMany
+    {
+        return $this->belongsToMany(Image::class);
+    }
+
+    /**
      * @return MorphToMany
      */
     public function tags(): MorphToMany
     {
         return $this->morphToMany(Tag::class, 'taggable');
+    }
+
+    /**
+     * @return BelongsToMany
+     */
+    public function videos(): BelongsToMany
+    {
+        return $this->belongsToMany(Video::class);
     }
 
     /**