+++ tests/dummy/database/migrations/2020_06_13_143800_create_post_and_video_tables.php
@@ -32,6 +32,7 @@ class CreatePostAndVideoTables extends Migration
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
+            $table->softDeletes();
             $table->unsignedBigInteger('author_id');
             $table->timestamp('published_at')->nullable();
             $table->string('title');
@@ -44,6 +45,12 @@ class CreatePostAndVideoTables extends Migration
                 ->on('users')
                 ->onDelete('cascade')
                 ->onUpdate('cascade');
+        });
+
+        Schema::create('images', function (Blueprint $table) {
+            $table->uuid('uuid')->primary();
+            $table->timestamps();
+            $table->string('url');
         });
 
         Schema::create('videos', function (Blueprint $table) {
@@ -100,6 +107,42 @@ class CreatePostAndVideoTables extends Migration
                 ->on('tags')
                 ->cascadeOnUpdate()
                 ->cascadeOnDelete();
+        });
+
+        Schema::create('image_post', function (Blueprint $table) {
+            $table->uuid('image_uuid');
+            $table->unsignedBigInteger('post_id');
+            $table->primary(['image_uuid', 'post_id']);
+
+            $table->foreign('image_uuid')
+                ->references('uuid')
+                ->on('images')
+                ->cascadeOnDelete()
+                ->cascadeOnUpdate();
+
+            $table->foreign('post_id')
+                ->references('id')
+                ->on('posts')
+                ->cascadeOnDelete()
+                ->cascadeOnUpdate();
+        });
+
+        Schema::create('post_video', function (Blueprint $table) {
+            $table->unsignedBigInteger('post_id');
+            $table->uuid('video_uuid');
+            $table->primary(['post_id', 'video_uuid']);
+
+            $table->foreign('post_id')
+                ->references('id')
+                ->on('posts')
+                ->cascadeOnDelete()
+                ->cascadeOnUpdate();
+
+            $table->foreign('video_uuid')
+                ->references('uuid')
+                ->on('videos')
+                ->cascadeOnDelete()
+                ->cascadeOnUpdate();
         });
     }