+++ tests/dummy/tests/Api/V1/Serializer.php
@@ -43,12 +43,13 @@ class Serializer
             'id' => (string) $post->getRouteKey(),
             'attributes' => [
                 'content' => $post->content,
-                'createdAt' => $post->created_at->jsonSerialize(),
+                'createdAt' => optional($post->created_at)->jsonSerialize(),
+                'deletedAt' => optional($post->deleted_at)->jsonSerialize(),
                 'publishedAt' => optional($post->published_at)->jsonSerialize(),
                 'slug' => $post->slug,
                 'synopsis' => $post->synopsis,
                 'title' => $post->title,
-                'updatedAt' => $post->updated_at->jsonSerialize(),
+                'updatedAt' => optional($post->updated_at)->jsonSerialize(),
             ],
             'relationships' => [
                 'author' => [
@@ -61,6 +62,12 @@ class Serializer
                     'links' => [
                         'self' => "{$self}/relationships/comments",
                         'related' => "{$self}/comments",
+                    ],
+                ],
+                'media' => [
+                    'links' => [
+                        'self' => "{$self}/relationships/media",
+                        'related' => "{$self}/media",
                     ],
                 ],
                 'tags' => [