+++ CHANGELOG.md
@@ -5,8 +5,12 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
+### Added
+
 - [#43](https://github.com/laravel-json-api/laravel/issues/43) The package now supports soft-deleting resources. For
   full details on how to apply this to resource schemas, refer to the new *Soft Deleting* chapter in the documentation.
+- Multi-resource models are now supported. This allows developers to represent a single model class as multiple
+  different JSON:API resource types within an API. Refer to documentation for details of how to implement.
 
 ## [1.0.0-alpha.4] - 2021-02-27